added remove item dialog

+++ src/app/cart/cart.component.ts
@@ -2,19 +2,28 @@ import { Component, OnInit } from "@angular/core";
 import { CartService } from "../cart.service";
 import { FormBuilder } from "@angular/forms";
 import { Product } from "../product";
-import { animate, state, style, transition, trigger } from "@angular/animations";
-import { MatDialog } from '@angular/material/dialog';
+import {
+  animate,
+  state,
+  style,
+  transition,
+  trigger
+} from "@angular/animations";
+import { MatDialog, MatDialogRef } from "@angular/material/dialog";
 
 @Component({
   selector: "app-cart",
   templateUrl: "./cart.component.html",
   styleUrls: ["./cart.component.css"],
   animations: [
-    trigger('detailExpand', [
-      state('collapsed', style({height: '0px', minHeight: '0'})),
-      state('expanded', style({height: '*'})),
-      transition('expanded <=> collapsed', animate('225ms cubic-bezier(0.4, 0.0, 0.2, 1)')),
-    ]),
+    trigger("detailExpand", [
+      state("collapsed", style({ height: "0px", minHeight: "0" })),
+      state("expanded", style({ height: "*" })),
+      transition(
+        "expanded <=> collapsed",
+        animate("225ms cubic-bezier(0.4, 0.0, 0.2, 1)")
+      )
+    ])
   ]
 })
 export class CartComponent implements OnInit {
@@ -46,13 +55,28 @@ export class CartComponent implements OnInit {
   }
 
   removeItem(product: Product) {
-    this.dialog.open(CartRemoveItemDialog);
-    this.cartService.removeItem(product);
+    const dialogRef = this.dialog.open(CartRemoveItemDialog);
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result.event == "yes") {
+        this.cartService.removeItem(product);
+      }
+    });
   }
 }
 
 @Component({
-  selector: 'cart-remove-item-dialog',
-  templateUrl: 'cart-remove-item-dialog.html',
+  selector: "cart-remove-item-dialog",
+  templateUrl: "cart-remove-item-dialog.html"
 })
-export class CartRemoveItemDialog {}
+export class CartRemoveItemDialog {
+  constructor(public dialogRef: MatDialogRef<CartRemoveItemDialog>) {}
+
+  onNo() {
+    this.dialogRef.close({ event: "no" });
+  }
+
+  onYes() {
+    this.dialogRef.close({ event: "yes" });
+  }
+}
